[Ruby] Fix memory leak in GRPC::Core::Call#run_batch
The C function that backs this Ruby method can raise a variety of
errors. In particular, grpc_run_batch_stack_fill_ops has a lot of places
that raise errors. When an error is raised, allocated memory (e.g.
run_batch_stack) is not freed and is permanently leaked.

This changes grpc_rb_call_run_batch to have an ensure block to perform
cleanup and free all allocated resources.