ruby: register grpc_rb_sStatus as a global variable
Ref: https://bugs.ruby-lang.org/issues/20311

C global variable that contain references to Ruby object MUST
be declared to the Ruby GC to prevent these objects from being
collected or moved.

There are a few exceptions to that, such as classes defined using
the C APIs such as `rb_define_class` etc.

Up to Ruby 3.4 however, there was a bug that caused classes created
from Ruby with the `Struct.new("Name")` API to also be marked as
immortal and immovable.

GRPC has been relying on this bug, which I fixed in Ruby 3.4, and
now GRPC is crashing when `Struct::Status` is moved around by the GC.