Ruby: Mark credential object in channel
The credentials object potentially holds a reference to a proc
when composed with CallCredentials.

We need the proc to live so that when requests build the credentials
the proc can still be invoked.  For the proc to live we need the
credentials object to live so it needs to be marked by the channel.

Co-authored-by: Ufuk Kayserilioglu <ufuk.kayserilioglu@shopify.com>
Co-authored-by: Peter Zhu <peter.zhu@shopify.com>
Co-authored-by: Kevin Menard <kevin.menard@shopify.com>